+++ config/profiling.exs
@@ -1,0 +1,8 @@
+use Mix.Config
+
+config :td_dd, :profiling,
+  profiling_import_schema: %{
+    external_id: :string,
+    value: :map
+  },
+  profiling_import_required: [:external_id]