+++ lib/td_dd/data_structures/data_structure.ex
@@ -10,6 +10,7 @@ defmodule TdDd.DataStructures.DataStructure do
   alias TdDd.DataStructures
   alias TdDd.DataStructures.DataStructure
   alias TdDd.DataStructures.DataStructureVersion
+  alias TdDd.DataStructures.Profile
   alias TdDd.Searchable
   alias TdDd.Systems.System
   alias TdDfLib.Format
@@ -19,6 +20,7 @@ defmodule TdDd.DataStructures.DataStructure do
   schema "data_structures" do
     belongs_to(:system, System, on_replace: :update)
     has_many(:versions, DataStructureVersion, on_delete: :delete_all)
+    has_one(:profile, Profile, on_delete: :delete_all)
 
     field(:confidential, :boolean)
     field(:df_content, :map)